todo: 经测试 json 无法被格式化

+++ packages/basic/index.js
@@ -53,7 +53,7 @@ module.exports = {
       parser: 'jsonc-eslint-parser',
       rules: {
         // 双引号
-        'quotes': ['error', 'double'],
+        'quotes': ['error', 'single'],
         // 键值用引号括起来
         'quote-props': ['error', 'always'],
         // 末尾不加逗号
